+++ package.json
@@ -88,6 +88,8 @@
     "@storybook/test": "^8.6.4",
     "@storybook/vue3": "^8.6.4",
     "@storybook/vue3-vite": "^8.6.4",
+    "@types/body-parser": "^1.19.6",
+    "@types/wait-on": "^5.3.4",
     "cross-env": "^7.0.3",
     "eslint-plugin-storybook": "^0.12.0",
     "prettier-plugin-sort-json": "^4.1.1",
