+++ tests/battery-settings.spec.ts
