+++ tests/fatal.spec.ts
