+++ tests/simulator/vite.config.ts
