+++ tests/smart-cost-only.spec.ts
