+++ tests/smart-cost.spec.ts
