+++ tests/statistics.spec.ts
