+++ tests/vehicle-error.spec.ts
