+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
-  "include": ["env.d.ts", "assets/js/**/*"],
+  "include": ["env.d.ts", "assets/js/**/*", "tests/**/*"],
   "compilerOptions": {
     "allowUnreachableCode": false,
     "allowUnusedLabels": false,
